backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
pdfminer.six
transformers
sentencepiece
nltk
